+++ packages/firefuel/test/src/firefuel_repository_test.dart
@@ -275,13 +275,13 @@ void main() {
         testRepository.orderBy([OrderBy(field: TestUser.fieldName)], limit: 1),
   );
 
-  RepositoryTestUtil.runTests<Chunk<TestUser>, TestUser>(
+  RepositoryTestUtil.runTests<FirefuelPage<TestUser>, TestUser>(
     methodName: 'paginate',
     mockCollection: MockCollection(),
     initHappyPath: (mockCollection) async {
       mockCollection.initialize(
         onPaginate: () {
-          return Chunk<TestUser>(orderBy: [OrderBy(field: TestUser.fieldName)]);
+          return const Chunk<TestUser, DocumentSnapshot<TestUser?>>();
         },
       );
     },
@@ -290,7 +290,7 @@ void main() {
     },
     methodCallback: (testRepository) {
       return testRepository.paginate(
-        Chunk<TestUser>(orderBy: [OrderBy(field: TestUser.fieldName)]),
+        FirefuelQuery(orderBy: [OrderBy(field: TestUser.fieldName)]),
       );
     },
   );